Refactor datamining report routes
