--- src/app/datamining/bulletins/canonical-map.tsx
@@ -1,1 +1,0 @@
-export { default } from "./sandbox/page";
