--- src/app/datamining/placeholder/page.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-export default function PlaceholderPage() {
-  return (
-    <div className="p-6 text-gray-400">
-      🚧 Em breve: novos relatórios aqui...
-    </div>
-  );
-}
